move extracted class into own file
